backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
drf-spectacular>=0.27
django-filter>=23.5
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
requests>=2.32
xmltodict>=0.13
